move section creation to disassembler_section
